requirements.txt
RPi.GPIO==0.7.0
pynput==1.7.3
